+++ common/src/main/java/nl/teamdiopside/infinitybuttons/compat/blocks/NeapolitanBlocks.java
@@ -6,7 +6,7 @@ import net.minecraft.world.level.block.SoundType;
 import net.minecraft.world.level.block.state.BlockBehaviour;
 import net.minecraft.world.level.material.MapColor;
 import nl.teamdiopside.infinitybuttons.block.faced4.SecretButton;
-import nl.teamdiopside.infinitybuttons.block.faced4.SecretButtonShape;
+import nl.teamdiopside.infinitybuttons.block.faced4.SecretButtonType;
 import nl.teamdiopside.infinitybuttons.compat.IBModdedBlocks;
 
 import static com.mojang.text2speech.Narrator.LOGGER;
@@ -19,7 +19,7 @@ public class NeapolitanBlocks extends IBModdedBlocks {
             .ignitedByLava().strength(2.0f, 3.0f).noOcclusion().sound(SoundType.WOOD).requiresCorrectToolForDrops();
 
     public static final RegistrySupplier<SecretButton> CHISELED_CHOCOLATE_BRICK_SECRET_BUTTON = INSTANCE.registerSecretButton("chiseled_chocolate_brick_secret_button",
-            SecretButtonShape.CHISELED_NETHER_BRICK, properties, ResourceLocation.fromNamespaceAndPath(NAMESPACE, "chiseled_chocolate_bricks"));
+            SecretButtonType.CHISELED_NETHER_BRICK, properties, ResourceLocation.fromNamespaceAndPath(NAMESPACE, "chiseled_chocolate_bricks"));
 
     private NeapolitanBlocks() {
         super(NAMESPACE);