+++ CHANGELOG.md
@@ -3,3 +3,6 @@
 
 ## Technical Changes
 - Buttons now use the vanilla POWERED blockstate instead of PRESSED
+
+## Content
+- Added (Chiseled) Myalite Brick Secret Buttons from Quark